channel: handle channels with pending data
This series fixes a bug where a packet received on the TLS connection
was delivered to the TUN device until a subsequent packet was
received.

The TLS channel reads each packet in two steps, first the framing
header to get the size and then the actual packet.  It returns EAGAIN
to the channel layer after the first read to indicate that it must be
called again to complete the packet. The channel layer issues a read
poll on the underlying TCP socket and calls the TLS read again when
triggered.  The TCP polls are level triggered, so this should work,
even if the data is already available.

Unfortunately, this doesn't always work with TLS.  OpenSSL has to
consume an TLS record before decrypting. So to decrypt the framing
header, it must have also read the body from the TCP socket and
decrypted it.  The TCP read poll issued by the channel layer isn't
triggered (until another packet arrives) because the body is already
consumed and buffered internally by OpenSSL.

The fix here is add a pending() method to the channel that returns
true if the channel has internally buffered data that can be
immediately read.  The channel layer issues an immediate read in this
case, rather than configuring a poll.