tcp: enable TCP keepalives
When no real traffic is being sent or received, enftun needs to
periodically poll the ENF to maintain the TLS tunnel. This is
necessary both to detect if the connection has failed, but also to
keep active any NAT entries along the path.

There is an open issue to implement this functionality directly into
enftun, by pinging the ENF router. As a quick solution until that
feature is complete, we enable the built-in TCP keepalive
functionality.

This series configures keepalive support on all enftun TCP sockets. It
sets the paramaters to send a keepalive every 5 minutes, if the
connect is idle.  If a response is not received, it will try again
every 10 seconds for up to 1 minute. If no responses are received
during that minute, it will consider the connection dead.  enftun will
then start attempting to reconnect.