ndp: fix route lifetime in RA route info options
The RA route info options "route lifetime" parameter is specified in
seconds, but we were setting it will a value specified in
milliseconds.

The lifetime is allowed to infinite (a value of 0xffffffff), which is
most appropriate for our use case anyway.

So fix this issue by setting the lifetime to infinite.