Support TCP or PSOCK
This adds support for the custom psock protocol on enftun. The default 
is to use TCP, but enftun can be built with psock using the cmake flag 
BUILD_PSOCK. Currently, both tcp and psock cannot be used together, and 
when psock is used support for monitoring the connection state is lost.