context: zero out the ipv6 string buffer before reading cert
The ipv6 string was not zeroed before re-reading the certificate
file. If the cert changed such that the new ipv6 string was shorter
(for example, if a run of zeros was replaced with a ::), then the ipv6
string buffer would contain the concatentation of the new address and
the remainder of the old address. Obviously, this is invalid and
incorrect.

Prevent this errror by zeroing the ipv6 string buffer before reading
the certificate file.