add support for OpenSSL 1.0.x versions
Some IoT gateways and other embedded products have OpenSSL 1.0.x, not
the OpenSSL 1.1.x required by enftun.  Upgrading the version of
OpenSSL on these platforms is non-trivial, if even possible. So to
support them, to adapt enftun to the older version of OpenSSL.

Our usage of the new APIs in 1.1.x is limited to convenience methods
for restricting the allowed version of TLS to 1.2.  This same behavior
can be enforced on older versions with slightly more verbose code.